requirements.txt
elevenlabs==1.8.0
python-dotenv==1.0.0
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
twilio>=8.10.0
openai>=1.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'
